Initial prototype implementation of a Puppet Ruby host process.
This is an initial implementation of a Ruby host process that can load Puppet
functions and resource types implemented in Ruby.

The host process offers two services over gRPC (HTTP/2 transport):

* `type`: a service responsible for describing Puppet resource types
  implemented in Ruby.
* `function`: a service responsible for describing and invoking Puppet
  functions implemented in Ruby.

This is used in conjunction with the `--ruby-host` option to the native
compiler which will enable loading resource types and functions written in
Ruby.